# doc/openapi.yaml
openapi: 3.0.3
info:
  title: HealthPorta Healthcare MRF API
  version: "1.0.0"
  description: |
    OpenAPI description for the HealthPorta Healthcare MRF API.  The API
    is versioned under `/api/v1` and served by a Sanic application.  All
    endpoints return JSON.  Unless noted otherwise, error responses use
    the `Error` schema.
servers:
  - url: http://localhost:8085/api/v1
paths:
...
  /import/issuer/{issuer_id}:
    get:
      summary: Fetch import information for a specific issuer
      parameters:
        - name: issuer_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Issuer metadata and import errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImporterIssuerDetail'
        '404':
          description: Issuer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /issuer/id/{issuer_id}:
    get:
      summary: Retrieve issuer details and plans
      parameters:
        - name: issuer_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Issuer data, import errors, and plans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuerDetail'
        '404':
          description: Issuer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /issuer/:
    get:
      summary: List issuers across all states
      responses:
        '200':
          description: Array of issuer summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IssuerSummary'
        '404':
          description: No issuers found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /issuer/state/{state}:
    get:
      summary: List issuers for a specific state
      parameters:
        - name: state
          in: path
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 2
      responses:
        '200':
          description: Array of issuer summaries for the state
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IssuerSummary'
        '404':
          description: No issuers found for the state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      additionalProperties: true
...
    TierDescriptor:
      type: object
      properties:
        tier_slug:
          type: string
          enum: [tier_1, tier_2, tier_3, tier_4, tier_5, tier_6, generic, preferred_generic, non_preferred_generic, brand, preferred_brand, non_preferred_brand, specialty, preferred_specialty, non_preferred_specialty, other, unknown]
        tier_label:
          type: string
      required: [tier_slug, tier_label]
      additionalProperties: false
    TierBreakdownEntry:
      allOf:
        - $ref: '#/components/schemas/TierDescriptor'
        - type: object
          properties:
            drug_count:
              type: integer
          required: [drug_count]
          additionalProperties: false
...
    ImporterIssuerDetail:
      type: object
      properties:
        issuer_id:
          type: integer
        issuer_name:
          type: string
        import_errors_count:
          type: integer
        err_log:
          type: array
          items:
            type: object
            additionalProperties: true
      additionalProperties: true
    IssuerSummary:
      type: object
      properties:
        issuer_id:
          type: integer
        issuer_name:
          type: string
        state:
          type: string
        import_errors:
          type: integer
        plan_count:
          type: integer
      additionalProperties: true
    IssuerDetail:
      allOf:
        - $ref: '#/components/schemas/IssuerSummary'
        - type: object
          properties:
            plans:
              type: array
              items:
                type: object
                additionalProperties: true
            drug_summary:
              $ref: '#/components/schemas/IssuerDrugSummary'
    IssuerDrugSummary:
      type: object
      properties:
        total_drugs:
          type: integer
        authorization:
          type: object
          properties:
            required:
              type: integer
            not_required:
              type: integer
          additionalProperties: false
        step_therapy:
          type: object
          properties:
            required:
              type: integer
            not_required:
              type: integer
          additionalProperties: false
        quantity_limits:
          type: object
          properties:
            has_limit:
              type: integer
            no_limit:
              type: integer
          additionalProperties: false
        tiers:
          type: array
          items:
            $ref: '#/components/schemas/TierBreakdownEntry'
      required: [total_drugs, authorization, step_therapy, quantity_limits, tiers]
      additionalProperties: false
...